requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.19.1
numpy==1.26.4
textblob==0.17.1
pyspellchecker==0.7.2
nltk==3.8.1
python-dotenv==1.0.0